1. The arena_base lower-right regression — fixed Root cause: the ANTIC 4K scan-boundary trap, exposed by my split. In the original single-file build the screen buffer landed at $7c3b, putting the 23-row play region at $7c63–$7ffb — ending just 5 bytes below $8000. My split marked the shared globals inline (external linkage), which grew .text by ~24 bytes and shifted the screen buffer to $7c53. That pushed the play region to $7c7b–$8013, so the bottom row straddled $8000 mid-row — ANTIC's 12-bit scan counter wraps at the 4K page, corrupting the right ~19 columns of that row (the missing lower-right). The display-list builder (display_list.h:112) can only insert an LMS reload at row starts, so a mid-row straddle is unfixable from the display list.
Fix: reverted the shared-header data/state from inline back to static/constexpr (internal linkage). Each arena demo is a single TU, so internal linkage is correct and it restores the original BSS layout. Result: the screen buffer is back at $7c3b, and both .xex files are now byte-identical to fresh builds of the pre-split source — provably behavior-preserving.

2. Renamed arena_base → arena_native
Both CMake build paths now use target/output arena_native → arena_native.xex (matching arena_native.cpp); arena_vbxe was already consistent. No arena_base references remain. I also removed the stale, untracked, now-orphaned demo/arena/arena_base.xex build artifact.

Final state — demo/arena/ holds arena_native.cpp, arena_vbxe.cpp, arena_shared.h; zero EDGE_VBXE in any of them; both targets build clean.

⚠️ One thing to keep in mind: that 5-byte margin to $8000 is a genuine latent fragility (it's in the project's ANTIC-4K memory) — any future change that grows arena_native's code/BSS can re-trigger this. The robust long-term fix would be engine-side: aligning the Core screen buffer (e.g. to 0x400) so a ≤1024-byte buffer can never straddle a 4K page. I didn't do that here since it touches the engine and affects all games — say the word if you'd like me to take that on. I also recorded the gotcha to memory for future sessions.
